requirements.txt
git+https://github.com/yt-dlp/yt-dlp
colorama==0.4.4
ffmpeg-python==0.2.0
spotipy==2.19.0
halo==0.0.31
numpy==1.23.3
git+https://github.com/sigma67/ytmusicapi